Refactor Callback into base callbacks
Mostly moved this around to eventually share callback initialization logic between train and optimize
