+++ rl_algo_impls/shared/callbacks/microrts_reward_decay_callback.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from rl_algo_impls.runner.config import Config
-from rl_algo_impls.shared.callbacks.callback import Callback
+from rl_algo_impls.shared.callbacks import Callback
 from rl_algo_impls.wrappers.vectorable_wrapper import VecEnv
